+++ airflow/cli/commands/kubernetes_command.py
@@ -25,9 +25,10 @@ from kubernetes import client
 from kubernetes.client.api_client import ApiClient
 from kubernetes.client.rest import ApiException
 
-from airflow.executors.kubernetes_executor import KubeConfig, create_pod_id
+from airflow.executors.kubernetes_executor import KubeConfig
 from airflow.kubernetes import pod_generator
 from airflow.kubernetes.kube_client import get_kube_client
+from airflow.kubernetes.kubernetes_helper_functions import create_pod_id
 from airflow.kubernetes.pod_generator import PodGenerator
 from airflow.models import DagRun, TaskInstance
 from airflow.utils import cli as cli_utils, yaml
